Update dynamic_detection.m
This program finds the best possible K value for any given signal. You can edit the variables in the begining such as SNR, Length of the scanned signal, Sampling frequency of the energy detector and the n point fft. Also the optimum value of K can be computed by fixing a value for Pf. the default value already set is 0.1